pyproject.toml
[tool.poetry]
name = "nonebot-plugin-smms"
version = "0.1.0"
description = "sm.ms 图床插件"
authors = ["mobyw <mobyw66@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/mobyw/nonebot-plugin-smms"
repository = "https://github.com/mobyw/nonebot-plugin-smms"
documentation = "https://github.com/mobyw/nonebot-plugin-smms/blob/main/README.md"
keywords = ["nonebot2", "qq", "plugin"]
packages = [
    { include = "nonebot_plugin_smms" }
]

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = {extras = ["httpx"], version = "^2.1.2"}

[tool.poetry.dev-dependencies]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 88
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
src_paths = ["plugins"]
extra_standard_library = ["typing_extensions"]

[tool.ruff]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "C901"]
line-length = 88
target-version = "py38"

[tool.ruff.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
